requirements.txt
-e .
aws_cdk.core
aws_cdk.aws_s3
aws_cdk.aws_glue
aws_cdk.aws_iam
aws_cdk.aws_athena
aws_cdk.aws_kinesisfirehose
aws_cdk.aws_kinesis
aws_cdk.aws_rds
aws_cdk.aws_ec2
aws_cdk.aws_dms
aws_cdk.aws_emr
aws_cdk.aws_redshift
black==20.8b1
pre-commit==2.9.3
fake-web-events
boto3
pandas
psycopg2-binary